requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.7
distlib==0.3.2
filelock==3.0.12
idna==3.3
jaraco.classes==3.2.1
MouseInfo==0.1.3
mss==6.1.0
multidict==5.2.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.3.0
psutil==5.8.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dotenv==0.19.1
pytweening==1.0.4
requests==2.26.0
selenium==3.141.0
six==1.16.0
twitchio==2.0.6
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.7.2
virtualenv-clone==0.5.7
yarl==1.7.0
